toolkit/devtools/server/tests/mochitest/test_inspector-remove.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
const inspector = require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gClient = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

function ignoreNode(node) {
  // Duplicate the walker logic to skip blank nodes...
  return node.nodeType === Components.interfaces.nsIDOMNode.TEXT_NODE &&
    !/[^\s]/.test(node.nodeValue);
}

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testRemoveSubtree() {
  let originalOwnershipSize = 0;
  let longlist = null;
  let longlistID = null;

  let nextSibling = gInspectee.querySelector("#longlist").nextSibling;
  while (nextSibling && ignoreNode(nextSibling)) {
    nextSibling = nextSibling.nextSibling;
  }

  let previousSibling = gInspectee.querySelector("#longlist").previousSibling;
  while (previousSibling && ignoreNode(previousSibling)) {
    previousSibling = previousSibling.previousSibling;
  }

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#longlist").then(listFront => {
    longlist = listFront;
    longlistID = longlist.actorID;
  }).then(() => {
    return gWalker.children(longlist);
  }).then((items)=> {
    originalOwnershipSize = assertOwnership();
    // Here is how the ownership tree is summed up:
    // #document                      1
    //   <html>                       1
    //     <body>                     1
    //       <div id=longlist>        1
    //         <div id=a>a</div>   26*2 (each child plus it's singleTextChild)
    //         ...
    //         <div id=z>z</div>
    //                             -----
    //                               56
    is(originalOwnershipSize, 56, "Correct number of items in ownership tree");
    return gWalker.removeNode(longlist);
  }).then(siblings => {
    is(siblings.previousSibling.rawNode(), previousSibling, "Should have returned the previous sibling.");
    is(siblings.nextSibling.rawNode(), nextSibling, "Should have returned the next sibling.");
    return waitForMutation(gWalker, isChildList);
  }).then(() => {
    // Our ownership size should now be 51 fewer (we forgot about #longlist + 26
    // children + 26 singleTextChild nodes, but learned about #longlist's
    // prev/next sibling)
    let newOwnershipSize = assertOwnership();
    is(newOwnershipSize, originalOwnershipSize - 51,
      "Ownership tree should be lower");
    // Now verify that some nodes have gone away
    return checkMissing(gClient, longlistID);
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gClient;
  runNextTest();
});


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
